requirements.txt
aiosmtplib>=2.0.0
python-telegram-bot>=20.0
ldap3>=2.9.0